refactor test code
